+++ lib/routes/app_router.gr.dart
@@ -9,17 +9,19 @@
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:auto_route/auto_route.dart' as _i12;
-import 'package:cloud_firestore/cloud_firestore.dart' as _i17;
 import 'package:flutter/material.dart' as _i13;
 import 'package:zachranobed/features/debug/debug_screen.dart' as _i3;
+import 'package:zachranobed/features/foodboxes/domain/model/box_movement.dart'
+    as _i14;
+import 'package:zachranobed/features/foodboxes/presentation/screen/box_movement_detail_screen.dart'
+    as _i1;
 import 'package:zachranobed/features/login/presentation/screen/login_screen.dart'
     as _i7;
-import 'package:zachranobed/models/box_movement.dart' as _i14;
-import 'package:zachranobed/models/offered_food.dart' as _i16;
-import 'package:zachranobed/models/user_data.dart' as _i15;
-import 'package:zachranobed/ui/screens/box_movement_detail_screen.dart' as _i1;
+import 'package:zachranobed/features/offeredfood/domain/model/offered_food.dart'
+    as _i15;
+import 'package:zachranobed/features/offeredfood/presentation/screens/donated_food_detail_screen.dart'
+    as _i4;
 import 'package:zachranobed/ui/screens/change_password_screen.dart' as _i2;
-import 'package:zachranobed/ui/screens/donated_food_detail_screen.dart' as _i4;
 import 'package:zachranobed/ui/screens/forgot_password_screen.dart' as _i5;
 import 'package:zachranobed/ui/screens/home_screen.dart' as _i6;
 import 'package:zachranobed/ui/screens/menu_screen.dart' as _i8;
@@ -40,7 +42,6 @@ abstract class $AppRouter extends _i12.RootStackRouter {
         child: _i1.BoxMovementDetailScreen(
           key: args.key,
           boxMovement: args.boxMovement,
-          user: args.user,
         ),
       );
     },
@@ -108,7 +109,7 @@ abstract class $AppRouter extends _i12.RootStackRouter {
         routeData: routeData,
         child: _i11.ThankYouScreen(
           key: args.key,
-          response: args.response,
+          isSuccess: args.isSuccess,
           message: args.message,
         ),
       );
@@ -123,14 +124,12 @@ class BoxMovementDetailRoute
   BoxMovementDetailRoute({
     _i13.Key? key,
     required _i14.BoxMovement boxMovement,
-    required _i15.UserData user,
     List<_i12.PageRouteInfo>? children,
   }) : super(
           BoxMovementDetailRoute.name,
           args: BoxMovementDetailRouteArgs(
             key: key,
             boxMovement: boxMovement,
-            user: user,
           ),
           initialChildren: children,
         );
@@ -145,18 +144,15 @@ class BoxMovementDetailRouteArgs {
   const BoxMovementDetailRouteArgs({
     this.key,
     required this.boxMovement,
-    required this.user,
   });
 
   final _i13.Key? key;
 
   final _i14.BoxMovement boxMovement;
 
-  final _i15.UserData user;
-
   @override
   String toString() {
-    return 'BoxMovementDetailRouteArgs{key: $key, boxMovement: $boxMovement, user: $user}';
+    return 'BoxMovementDetailRouteArgs{key: $key, boxMovement: $boxMovement}';
   }
 }
 
@@ -194,7 +190,7 @@ class DonatedFoodDetailRoute
     extends _i12.PageRouteInfo<DonatedFoodDetailRouteArgs> {
   DonatedFoodDetailRoute({
     _i13.Key? key,
-    required _i16.OfferedFood offeredFood,
+    required _i15.OfferedFood offeredFood,
     List<_i12.PageRouteInfo>? children,
   }) : super(
           DonatedFoodDetailRoute.name,
@@ -219,7 +215,7 @@ class DonatedFoodDetailRouteArgs {
 
   final _i13.Key? key;
 
-  final _i16.OfferedFood offeredFood;
+  final _i15.OfferedFood offeredFood;
 
   @override
   String toString() {
@@ -316,14 +312,14 @@ class OrderShippingOfBoxesRoute extends _i12.PageRouteInfo<void> {
 class ThankYouRoute extends _i12.PageRouteInfo<ThankYouRouteArgs> {
   ThankYouRoute({
     _i13.Key? key,
-    required _i17.DocumentReference<Object>? response,
+    required bool isSuccess,
     required String message,
     List<_i12.PageRouteInfo>? children,
   }) : super(
           ThankYouRoute.name,
           args: ThankYouRouteArgs(
             key: key,
-            response: response,
+            isSuccess: isSuccess,
             message: message,
           ),
           initialChildren: children,
@@ -338,18 +334,18 @@ class ThankYouRoute extends _i12.PageRouteInfo<ThankYouRouteArgs> {
 class ThankYouRouteArgs {
   const ThankYouRouteArgs({
     this.key,
-    required this.response,
+    required this.isSuccess,
     required this.message,
   });
 
   final _i13.Key? key;
 
-  final _i17.DocumentReference<Object>? response;
+  final bool isSuccess;
 
   final String message;
 
   @override
   String toString() {
-    return 'ThankYouRouteArgs{key: $key, response: $response, message: $message}';
+    return 'ThankYouRouteArgs{key: $key, isSuccess: $isSuccess, message: $message}';
   }
 }