+++ CHANGELOG.md
@@ -10,6 +10,14 @@ This is a changelog for **ZachranObed** application.
 ### Changed
 - **ZOB-105** Add non-prod firebase projects.
 - **ZOB-92** Update toolbar in Profile screen.
+- **ZOB-105** Add non-prod firebase projects.
+- **ZOB-114** Fetch user data from entities and entity pairs collections.
+- **ZOB-125** Change FCM tokens location in Firestore.
+- **ZOB-126** Fetch data for dashboard cards.
+- **ZOB-129** Fetch data for food boxes table.
+- **ZOB-133** Fetch data for offered food.
+- **ZOB-134** Write data to deliveries collection.
+- **ZOB-135** Fetch data for box movements.
 - **ZOB-150** Update pickup config time from 35min to 45min.
 
 ### Fixed